+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - update to the Msegat class structure.
 - add `log` config that allow you to send sms to log directly to use it you should add it to `env`
   as `MSEGAT_LOG = true`.
+- relocate `unicode` and `sender` from `Msegat` class into `MsegatMessage` object.